upd: code review fixes
