+++ src/Message/PurchaseRequest.php
@@ -70,7 +70,6 @@ class PurchaseRequest extends AuthorizeRequest
     {
         $data = parent::getData();
         $data['capture'] = 'true';
-        $data['confirmation_method'] = 'automatic';
 
         return $data;
     }
